requirements.txt
# Core Flask dependencies
Flask>=2.3.0
Werkzeug>=2.3.0
Jinja2>=3.1.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Audio/Video processing (Python 3.12 compatible)
moviepy>=1.0.3
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.26.0  # Python 3.12 compatible

# Image processing
Pillow>=10.0.0

# Google Cloud (optional)
google-cloud-storage>=2.10.0

# Additional audio dependencies
ffmpeg-python>=0.2.0

# WSGI server for production
gunicorn>=21.0.0

# Audio format support
pydub>=0.25.1

# Scientific computing (Python 3.12 compatible)
scipy>=1.11.0